+++ api_types.go
@@ -14,7 +14,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dgraph-io/dgraph/v24/x"
+	"github.com/hypermodeinc/dgraph/v24/x"
 	"github.com/hypermodeinc/modusdb/api/apiutils"
 	"github.com/hypermodeinc/modusdb/api/querygen"
 )