+++ unit_test/vector_test.go
@@ -17,7 +17,7 @@ import (
 	"testing"
 
 	"github.com/dgraph-io/dgo/v240/protos/api"
-	"github.com/dgraph-io/dgraph/v24/dgraphapi"
+	"github.com/hypermodeinc/dgraph/v24/dgraphapi"
 	"github.com/stretchr/testify/require"
 
 	"github.com/hypermodeinc/modusdb"