+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - feat: update interface to engine and namespace
   [#57](https://github.com/hypermodeinc/modusDB/pull/57)
 
+- chore: Update dgraph dependency [#62](https://github.com/hypermodeinc/modusDB/pull/62)
+
 ## 2025-01-02 - Version 0.1.0
 
 Baseline for the changelog.